+++ src/lib/components/nodes/EmptyConstructor.svelte
@@ -34,7 +34,9 @@
 
 <div class="custom-node">
   <div class="label">{data.type}</div>
-  <Handle id="output-self" type="source" position={Position.Right} />
+  {#each data.outputs as i, index (i)}
+    <Handle id={`output-${index}`} type="source" position={Position.Right} />
+  {/each}
 </div>
 
 <style>